[*] int -> size_t

+++ boot/ml/src/source/opencv/OpenCVFace.cpp
@@ -12,7 +12,7 @@ void lifuren::face(const std::string& model, const std::string& path) {
     }
     std::vector<cv::Rect> vector;
     classifier.detectMultiScale(image, vector);
-    for(int i = 0; i < vector.size(); ++i) {
+    for(size_t i = 0; i < vector.size(); ++i) {
         cv::rectangle(image, vector[i].tl(), vector[i].br(), cv::Scalar(255, 0, 255), 3);
     }
     cv::imshow("FaceImage", image);
